StationLocator/VAT-Spy: use Sets for coordinates
As the same coordinates are encountered multiple times for the same
object the resulting average was shifted/biased towards the majority.
Because that multiplicity is mostly an effect of technical limitations
of the format the resulting bias is usually unwanted.

Geo coordinates are now deduplicated before averaging to avoid such
shifts.