updated Log4j from 2.17.0 to 2.17.1 (CVE-2021-44832, unaffected)
Log4j 2.17.1 fixes CVE-2021-44832 which is a continuation of recent CVEs.

Unless users modify logging configuration files, this project is unaffected
by the reported vulnerability.

META-INF/NOTICE currently holds an invalid copyright year. NOTICE files
need to be reproduced by downstream software so it has to be included
uncorrected. An additional notice was added on top of the original copy.
The issue has been reported to upstream as LOG4J2-3375.